Improve prompting
Add best and last kernels to prompt.

This avoids an issue where a model generates a best kernel at some
iteration, then gets stuck trying to improve it but receiving no
feedback on compilation or correctness feedback on subsequent
generations, as it is always only shown the best kernel, and not the
last kernel it generated.